tests/requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
PyYAML>=6.0
requests>=2.28.0
tomli>=2.0.0
